# hw12_13_14_15_16_calendar/api/gen/go/api_event.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Swagger calendar - OpenAPI 3.0
 *
 * Swager calendar event
 *
 * API version: 1.0.12
 */

package api

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// EventAPIController binds http requests to an api service and writes the service results to the http response
type EventAPIController struct {
	service EventAPIServicer
	errorHandler ErrorHandler
}

// EventAPIOption for how the controller is set up.
type EventAPIOption func(*EventAPIController)

// WithEventAPIErrorHandler inject ErrorHandler into controller
func WithEventAPIErrorHandler(h ErrorHandler) EventAPIOption {
	return func(c *EventAPIController) {
		c.errorHandler = h
	}
}

// NewEventAPIController creates a default api controller
func NewEventAPIController(s EventAPIServicer, opts ...EventAPIOption) *EventAPIController {
	controller := &EventAPIController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the EventAPIController
func (c *EventAPIController) Routes() Routes {
	return Routes{
		"Updateevent": Route{
			"Updateevent",
			strings.ToUpper("Put"),
			"/event/{eventId}",
			c.Updateevent,
		},
		"Deleteevent": Route{
			"Deleteevent",
			strings.ToUpper("Delete"),
			"/event/{eventId}",
			c.Deleteevent,
		},
		"GetAllEvents": Route{
			"GetAllEvents",
			strings.ToUpper("Get"),
			"/events",
			c.GetAllEvents,
		},
		"Addevent": Route{
			"Addevent",
			strings.ToUpper("Post"),
			"/events",
			c.Addevent,
		},
	}
}

// OrderedRoutes returns all the api routes in a deterministic order for the EventAPIController
func (c *EventAPIController) OrderedRoutes() []Route {
	return []Route{
		Route{
			"Updateevent",
			strings.ToUpper("Put"),
			"/event/{eventId}",
			c.Updateevent,
		},
		Route{
			"Deleteevent",
			strings.ToUpper("Delete"),
			"/event/{eventId}",
			c.Deleteevent,
		},
		Route{
			"GetAllEvents",
			strings.ToUpper("Get"),
			"/events",
			c.GetAllEvents,
		},
		Route{
			"Addevent",
			strings.ToUpper("Post"),
			"/events",
			c.Addevent,
		},
	}
}



// Updateevent - Update event.
func (c *EventAPIController) Updateevent(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	eventIdParam, err := parseNumericParameter[int64](
		params["eventId"],
		WithRequire[int64](parseInt64),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "eventId", Err: err}, nil)
		return
	}
	var eventParam Event
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&eventParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertEventRequired(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertEventConstraints(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.Updateevent(r.Context(), eventIdParam, eventParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// Deleteevent - Deletes a event.
func (c *EventAPIController) Deleteevent(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	eventIdParam, err := parseNumericParameter[int64](
		params["eventId"],
		WithRequire[int64](parseInt64),
	)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Param: "eventId", Err: err}, nil)
		return
	}
	result, err := c.service.Deleteevent(r.Context(), eventIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// GetAllEvents - Read all events.
func (c *EventAPIController) GetAllEvents(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetAllEvents(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}

// Addevent - Add event event.
func (c *EventAPIController) Addevent(w http.ResponseWriter, r *http.Request) {
	var eventParam Event
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&eventParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertEventRequired(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	if err := AssertEventConstraints(eventParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.Addevent(r.Context(), eventParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	_ = EncodeJSONResponse(result.Body, &result.Code, w)
}